requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
cached-property==1.5.1
cffi==1.12.3
cloudpickle==1.2.2
constantly==15.1.0
cryptography==2.7
cssselect==1.0.3
easygui==0.98.1
future==0.18.2
gym==0.15.4
haralyzer==1.6.0
harparser==0.4
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
lxml==4.3.3
neat-python==0.92
numpy==1.17.3
opencv-python==4.1.1.26
parsel==1.5.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyDispatcher==2.0.5
pygame==1.9.6
pyglet==1.3.2
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
python-dateutil==2.8.0
queuelib==1.5.0
scipy==1.3.2
Scrapy==1.6.0
scrapy-splash==0.7.2
service-identity==18.1.0
six==1.12.0
Twisted==19.2.0
url-normalize==1.4.1
virtualenv==16.6.0
w3lib==1.20.0
zope.interface==4.6.0
